Rework metadata and generics
