+++ crates/rune-core/src/type_of.rs
@@ -1,0 +1,8 @@
+use crate::hash::Hash;
+
+/// Full type information.
+#[derive(Debug, Clone)]
+pub struct FullTypeOf {
+    #[doc(hidden)]
+    pub hash: Hash,
+}
