+++ crates/rune/Cargo.toml
@@ -18,7 +18,7 @@ default = ["emit", "std"]
 emit = ["codespan-reporting"]
 bench = []
 workspace = ["toml", "semver", "relative-path", "serde-hashkey", "linked-hash-map"]
-doc = ["rust-embed", "handlebars", "pulldown-cmark", "syntect", "sha2", "base64"]
+doc = ["rust-embed", "handlebars", "pulldown-cmark", "syntect", "sha2", "base64", "rune-core/doc"]
 cli = ["doc", "bincode", "atty", "tracing-subscriber", "clap", "webbrowser", "capture-io", "disable-io", "languageserver", "fmt", "similar"]
 languageserver = ["lsp", "ropey", "percent-encoding", "url", "serde_json", "tokio", "tokio/macros", "tokio/io-std", "workspace", "doc"]
 capture-io = ["parking_lot"]
@@ -29,8 +29,8 @@ alloc = []
 
 [dependencies]
 rune-macros = { version = "=0.12.3", path = "../rune-macros" }
+rune-core = { version = "=0.12.3", path = "../rune-core" }
 
-byteorder = { version = "1.4.3", default-features = false }
 futures-core = { version = "0.3.28", default-features = false }
 futures-util = { version = "0.3.28", default-features = false, features = ["alloc"] }
 itoa = "1.0.6"
@@ -42,7 +42,6 @@ serde_bytes = { version = "0.11.9", default-features = false, features = ["alloc
 smallvec = { version = "1.10.0", default-features = false, features = ["serde", "const_new"] }
 thiserror-impl = { version = "1.0.40", default-features = false }
 tracing =  { version = "0.1.37", default-features = false, features = ["attributes"] }
-twox-hash = { version = "1.6.3", default-features = false }
 hashbrown = { version = "0.13.2", features = ["serde"] }
 
 anyhow = { version = "1.0.71", features = ["std"], optional = true }
