+++ crates/rune/src/languageserver/completion.rs
@@ -107,7 +107,7 @@ pub(super) fn complete_native_instance_data(
                 .1
                 .return_type
                 .and_then(|hash| context.lookup_meta_by_hash(hash).first())
-                .map(|r| r.item.clone());
+                .and_then(|r| r.item.as_deref());
 
             let docs = meta.map(|meta| meta.docs.lines().join("\n"));
             let args = meta
@@ -169,7 +169,7 @@ pub(super) fn complete_native_loose_data(
                 .1
                 .return_type
                 .and_then(|hash| context.lookup_meta_by_hash(hash).first())
-                .map(|r| r.item.clone());
+                .and_then(|r| r.item.as_deref());
 
             let docs = meta.map(|meta| meta.docs.lines().join("\n"));
             let args = meta
