+++ crates/rune/src/params.rs
@@ -1,33 +1,9 @@
-use crate::hash::{Hash, IntoHash};
+use crate::hash::Hash;
 use crate::module::{AssociatedFunctionName, ToFieldFunction, ToInstance};
-use crate::runtime::{FullTypeOf, Protocol};
+use crate::runtime::Protocol;
 
-/// Helper to register a parameterized function.
-///
-/// This is used to wrap the name of the function in order to associated
-/// parameters with it.
-#[derive(Clone)]
-pub struct Params<T, const N: usize> {
-    pub(crate) name: T,
-    pub(crate) parameters: [FullTypeOf; N],
-}
-
-impl<T, const N: usize> Params<T, N> {
-    /// Construct a new parameters wrapper.
-    pub const fn new(name: T, parameters: [FullTypeOf; N]) -> Self {
-        Self { name, parameters }
-    }
-}
-
-impl<T, const N: usize> IntoHash for Params<T, N>
-where
-    T: IntoHash,
-{
-    #[inline]
-    fn into_hash(self) -> Hash {
-        self.name.into_hash()
-    }
-}
+#[doc(inline)]
+pub use rune_core::Params;
 
 impl<T, const N: usize> ToInstance for Params<T, N>
 where
